batch: Move gitlink storage
Batch storage now keeps text persistence plus the submodule pointer batch
path, while binary persistence lives in its own module.

Gitlink batch publication has atomic metadata and tree update rules
separate from text storage. Keeping it in storage leaves the module
carrying another file-type-specific operation.

This commit completes separating atomic file implementations by moving
gitlink batch persistence into batch.gitlink_storage and updating the
production caller.

Subsequent commits will move architecture checks to the new gitlink owner.
