+++ src/git_stage_batch/commands/selection/whole_file_batch_staging.py
@@ -5,10 +5,10 @@ from __future__ import annotations
 import sys
 
 from ...batch.binary_file_storage import add_binary_file_to_batch
+from ...batch.gitlink_storage import add_gitlink_to_batch
 from ...batch.ownership import BatchOwnership
 from ...batch.storage import (
     add_file_to_batch,
-    add_gitlink_to_batch,
 )
 from ...core.hashing import (
     compute_binary_file_hash,
